+++ example/src/screens/Casting.tsx
@@ -15,8 +15,6 @@ function prettyPrint(header: string, obj: any) {
 }
 
 export default function Casting() {
-  BitmovinCastManager.initialize();
-
   if (Platform.OS === 'android') {
     // Must be called in every activity on Android
     BitmovinCastManager.updateContext();